+++ lag_llama/gluon/lightning_module.py
@@ -12,9 +12,9 @@ from gluonts.itertools import prod
 from gluonts.torch.modules.loss import DistributionLoss, NegativeLogLikelihood
 from gluonts.torch.util import repeat_along_dim, take_last
 
-from data.augmentations.freq_mask import freq_mask
-from data.augmentations.freq_mix import freq_mix
-from data.augmentations.gluonts_augmentations import (
+from ...data.augmentations.freq_mask import freq_mask
+from ...data.augmentations.freq_mix import freq_mix
+from ...data.augmentations.gluonts_augmentations import (
     ApplyAugmentations,
     Jitter,
     MagnitudeWarp,
@@ -25,10 +25,10 @@ from data.augmentations.gluonts_augmentations import (
     WindowSlice,
     WindowWarp,
 )
-from gluon_utils.gluon_ts_distributions.implicit_quantile_network import (
+from ...gluon_utils.gluon_ts_distributions.implicit_quantile_network import (
     ImplicitQuantileNetworkOutput,
 )
-from lag_llama.model.module import LagLlamaModel
+from ..model.module import LagLlamaModel
 
 
 class LagLlamaLightningModule(LightningModule):
@@ -94,7 +94,7 @@ class LagLlamaLightningModule(LightningModule):
         self.save_hyperparameters()
         self.context_length = self.hparams.context_length
         self.prediction_length = self.hparams.prediction_length
-        self.model = LagLlamaModel(**self.hparams.model_kwargs)
+        self.model = LagLlamaModel(**self.hparams.model_kwargs).double()
         self.loss = self.hparams.loss
         self.lr = self.hparams.lr
         self.weight_decay = self.hparams.weight_decay
